Remove superfluous cast
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@19560 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/snmplib/snmp_transport.c
@@ -286,8 +286,8 @@ netsnmp_transport_recv(netsnmp_transport *t, void *packet, int length,
                                                   opaque ? *opaque : NULL,
                                                   olength ? *olength : 0);
         if (debugLength)
-            DEBUGMSGT_NC(("transport:recv","%lu bytes from %s\n",
-                          (unsigned long)length, str));
+            DEBUGMSGT_NC(("transport:recv","%d bytes from %s\n",
+                          length, str));
         SNMP_FREE(str);
     }
